Makefile: do not fail if $INITRD_DIR alredy exists
When building locally the $INITRD_DIR might be already present and possibly contains files. 
This commit prevents mkdir to fail, when trying to create the existing directory again. It also stops rm to fail on removing the directory by just deleting the containing files.
These problems will never be seen on the buildbots, as they are starting on an empty work-directory.